% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgePensionRatefn.R
\name{AgePensionRatefn}
\alias{AgePensionRatefn}
\title{Age Pension Rates Function}
\usage{
AgePensionRatefn(HHtbl = NULL)
}
\arguments{
\item{df}{Argument is HHtbl from the master age pension function}
}
\value{
A data frame
}
\description{
(Jackson to complete...)
}

