% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssessableAssetsfn.R, R/DeemedAssetsfn.R
\name{AssessableAssetsfn}
\alias{AssessableAssetsfn}
\alias{DeemedAssetsfn}
\title{Age Pension AssessableAssetsfn}
\usage{
AssessableAssetsfn(HHtbl = NULL)

DeemedAssetsfn(HHtbl = NULL)
}
\arguments{
\item{df}{Argument is HHtbl from the master age pension function}

\item{df}{Argument is HHtbl from the master age pension function}
}
\value{
A data frame

A data frame
}
\description{
(Jackson to complete)

(Jackson to complete...)
}

