% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ysi_income_tax.R
\name{IMPfn}
\alias{IMPfn}
\title{Imputation Credits Function}
\usage{
IMPfn(ShareDiv = ShareDiv, Citizenship = Citizenship,
  InvestInc_p = InvestInc_p)
}
\value{
A list
}
\description{
Imputation Credits function for use on HIDLA, as part of the Income Tax Assesment
}

