% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ysi_income_tax.R
\name{income_tax_model_func}
\alias{income_tax_model_func}
\title{YSI Applied Income Tax Rates Function for modeling new rates and/or thresholds}
\usage{
income_tax_model_func(df = Null, Model_name = NULL)
}
\arguments{
\item{Model_name}{The the model name .}

\item{income}{A numeric vector of income ($)}
}
\value{
A vector
}
\description{
Can be used on taxable income (TaxInc) or another income source that uses marginal income tax rates.
}

