% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolkit.R
\name{inspect_ysi}
\alias{inspect_ysi}
\title{Sample view of a interesing cohort in set}
\usage{
inspect_ysi(df = NULL)
}
\arguments{
\item{df}{A dataframe of interest}
}
\value{
A data frame
}
\description{
Sample view of a interesing cohort in set
}

