% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ysi_income_tax.R
\name{MATOfn}
\alias{MATOfn}
\title{Mature Age Tax Offset (MATO) Function}
\usage{
MATOfn(WageIncImp_SalSac = WageIncImp_SalSac, YoB = YoB)
}
\arguments{
\item{WageIncImp_SalSac}{Whole dollar value of Wage income with salary sacrifice}

\item{YoB}{The year of birth}
}
\value{
A vector
}
\description{
Mature Age Tax Offset (MATO) for use on HIDLA, as part of the Income Tax Assesment
For use with mapply()
}

