% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ysi_income_tax.R
\name{MLfn}
\alias{MLfn}
\title{Medicare Levy Function}
\usage{
MLfn(Taxinc = NULL)
}
\arguments{
\item{TaxInc}{Whole dollar value of taxable income}
}
\value{
A list
}
\description{
Medicare Levy  for use on HIDLA, as part of the Income Tax Assesment
For use with sapply()
}
\details{
Requires TaxInc (Taxable Income) variable to have been calulated previously
}

