% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_variables.R
\name{na_to_zero}
\alias{na_to_zero}
\title{Convert na's to zeros}
\usage{
na_to_zero(input = NULL)
}
\arguments{
\item{input}{Typically a vector or variable}
}
\value{
A vector
}
\description{
For use with mutate_each_()
}

