% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrdinaryIncomefn.R
\name{OrdinaryIncomefn}
\alias{OrdinaryIncomefn}
\title{Age Pension Assessable income function}
\usage{
OrdinaryIncomefn(HHtbl = NULL)
}
\arguments{
\item{df}{Argument is HHtbl from the master age pension function}
}
\value{
A data frame
}
\description{
(Jackson to complete...)
}

