% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Property_ownership_functions.R
\name{OwnHomeAPfn}
\alias{OwnHomeAPfn}
\title{Age Pension Home ownership function}
\usage{
OwnHomeAPfn(HHtbl = NULL)
}
\arguments{
\item{df}{Argument is HHtbl from the master age pension function}
}
\value{
A data frame
}
\description{
(Jackson to complete...)
}

