% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ysi_income_tax.R
\name{RegIncTAXfn}
\alias{RegIncTAXfn}
\title{REG Income Tax Function}
\usage{
RegIncTAXfn(income = NULL, brackets = c(18200, 37000, 80000, 180000, Inf),
  rates = c(0, 0.19, 0.325, 0.37, 0.45))
}
\arguments{
\item{income}{Taxable Income}

\item{brackets}{Tax brackes as vector, typically with the last bracket being Inf}

\item{rates}{tax rates as vector, must equal number of brakets}
}
\value{
A vector
}
\description{
This function finds the amount of income tax payable, with the 2014 rates
}

