% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RentAssistfn.R
\name{RentAssistfn}
\alias{RentAssistfn}
\title{Age Pension Rent assistance eligibility function}
\usage{
RentAssistfn(HHtbl = NULL)
}
\arguments{
\item{df}{Argument is HHtbl from the master age pension function}
}
\value{
A data frame
}
\description{
(Jackson to complete...)
}

