% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SUPERTAXfn.R
\name{SUPERTAXfn}
\alias{SUPERTAXfn}
\title{Super Tax part of Income Tax}
\usage{
SUPERTAXfn(Super_Inc, Age)
}
\arguments{
\item{Super_Inc}{Whole dollar value of super income}

\item{Age}{An interger of the age of the tax payer}
}
\value{
A list
}
\description{
Super Tax Based on Super income and age, for apply to HILDA for Income tax calculation
}

