% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolkit.R
\name{write_master_names}
\alias{write_master_names}
\title{YSI Writing Master Variable Name Sheet}
\usage{
write_master_names(df = NULL)
}
\arguments{
\item{df}{Argument is intended to be the REVIEWED updated master table ready to write to file}
}
\value{
A data frame
}
\description{
Write your updated Master to file. Be careful to check that you are happy, and that you're not destroying everything!
}

