% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ysi_age_pension.R
\name{ysi_age_pension}
\alias{ysi_age_pension}
\title{Age Pension Master Function}
\usage{
ysi_age_pension(df = NULL)
}
\arguments{
\item{df}{Argument is intended to be a survey with the Tax_Matrix set, as a data frame}
}
\value{
A data frame
}
\description{
(Jackson to complete...)
}

