% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ysi_age_pensionOwnHome.R
\name{ysi_age_pensionOwnHome}
\alias{ysi_age_pensionOwnHome}
\title{Age Pension Master Function including the family home in assets test}
\usage{
ysi_age_pensionOwnHome(HHtbl = HHtbl)
}
\arguments{
\item{df}{Argument is HHtbl from the master age pension function}
}
\value{
A data frame
}
\description{
(Jackson to complete...)
}

