% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_variables.R
\name{ysi_family_mut}
\alias{ysi_family_mut}
\title{Family Status Variables}
\usage{
ysi_family_mut(df = NULL)
}
\arguments{
\item{df}{Argument is intended to be a fresh HILDA data frame}
}
\value{
A data frame
}
\description{
This function mutates variables for coupled, Non dependents over 25, Number coupled, Partnered and Single.
(Coupled,NonDepOver25, NoCoupled, Partnered, Single)
}

