% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ysi_income_tax.R
\name{ysi_income_tax}
\alias{ysi_income_tax}
\title{YSI Income Tax Assesment Function}
\usage{
ysi_income_tax(df = NULL)
}
\arguments{
\item{df}{Argument is intended to be a fresh HILDA data frame}
}
\value{
A data frame
}
\description{
This function combines all the nessesary income tax functions and data set mutations to be an all-in-one function for calculating Income Tax
}

