% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ysi_init.R
\name{ysi_init}
\alias{ysi_init}
\title{YSI Startup Function}
\usage{
ysi_init(set = "key", xwave = F, wave_n = 14)
}
\arguments{
\item{set}{This argument needs to be a string that is one of the set names in the YSI master sheet.}

\item{xwave}{Expects a logical input, xwave data (T) or single wave data (F)}

\item{wave_n}{If you want a single wave of data, specify which year in the form of an integer between 1 and 14}
}
\value{
A set of tables, including Master (the variable master sheet) and your_data (your data subset), and user_subset (filtered version of Master with your set's variables)
}
\description{
This function is for the star of your script, bringing in the data and setting the data up for your analysis.
There is no need to be concerned that your varible needs are not in every wave, as this function will just find whatever is avilable.
}

