% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_variables.R
\name{ysi_mutate_combine}
\alias{ysi_mutate_combine}
\title{ysi Mutate_combine Function}
\usage{
ysi_mutate_combine(df = NULL)
}
\arguments{
\item{df}{Argument is intended to be a data frame}
}
\value{
A data frame
}
\description{
This function mutates multiple useful and commonly used variables for YSI analysis.
}
\details{
It generates the following, which are mostly a sum of the posative and negative HILDA variables:

House Income (HouseInc)
Total Individual Income (TotalIncome)
Regular Individual Income (RegInc)
HouseHold level combined Assets (Assets)
Idividual Net Rent (NetRent)
}

