% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ysi_super_tax.R
\name{ysi_super_tax}
\alias{ysi_super_tax}
\title{YSI Super Tax Assesment Function (Working...)}
\usage{
ysi_super_tax(df = NULL)
}
\arguments{
\item{df}{Argument is intended to be a survey with the Tax_Matrix set, as a data frame}
}
\value{
A data frame
}
\description{
This function combines all the nessesary Superannuation tax functions and data set mutations to be an all-in-one function for calculating Super Tax
(Grace to fill in soon)
(Requires Tax_Matrix set)
(Currently only usable in Wave14)
}

