% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ysi_income_tax.R
\name{ysi_tax_inc}
\alias{ysi_tax_inc}
\title{YSI Taxable Income Standalone Function}
\usage{
ysi_tax_inc(df = NULL)
}
\arguments{
\item{df}{Argument is intended to be a fresh HILDA data frame}
}
\value{
A data frame
}
\description{
This function just finds taxable income.
}

