% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ysi_income_tax.R
\name{income_tax_rates_func}
\alias{income_tax_rates_func}
\title{YSI Applied Income Tax Rates Function}
\usage{
income_tax_rates_func(income = NULL, year = NULL)
}
\arguments{
\item{income}{A numeric vector of income ($)}

\item{year}{The finanical year of interest, in the form "2013-14", as a vector.}
}
\value{
A vector
}
\description{
Can be used on taxable income (TaxableInc) or another income source that uses marginal income tax rates.
Previously called RegIncTAXfn()
}
\details{
To model a new set of rates and thresholds use income_tax_model_func()
}

