% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolkit.R
\name{inspect_ysi}
\alias{add_more_vars}
\alias{inspect_ysi}
\title{Sample view of a interesing cohort in set}
\usage{
inspect_ysi(df = NULL)

add_more_vars(new_vars = NULL, set = NULL, name = T, new_names = NULL,
  wave_id = F)
}
\arguments{
\item{df}{A dataframe of interest}

\item{new_vars}{Argument is intended to be a survey with the Tax_Matrix set, as a data frame}

\item{set}{Argument is intended to be a survey with the Tax_Matrix set, as a data frame}

\item{new_names}{If the variables that you are adding are unnamed put the new names here}

\item{wave_id}{True if input has the wave letter as a prefix}

\item{names_bol}{True if input is the variable names, False if adding variables using the variable code}
}
\value{
A data frame

A data frame of the new Master
}
\description{
Sample view of a interesing cohort in set

YSI Editing Master Variable Name Sheet
}

