% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ysi_income_tax.R
\name{LITOfn}
\alias{LITOfn}
\title{Low Income Tax Offset (LITO) Function}
\usage{
LITOfn(TaxInc = TaxInc)
}
\arguments{
\item{TaxInc}{Whole dollar value of taxable income}
}
\value{
A vector
}
\description{
Low Income Tax Offset (LITO) for use on HIDLA, as part of the Income Tax Assesment
For use with sapply()
}
\details{
Requires TaxInc (Taxable Income) variable to have been calulated previously
}

