% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SUPERTAXfn.R
\name{SUPERTAXfn}
\alias{SUPERTAXfn}
\title{Super Tax Function}
\usage{
SUPERTAXfn(Super_Inc, Age)
}
\arguments{
\item{Super_Inc}{Whole dollar value of super income}

\item{Age}{The age of the tax payer}
}
\value{
A list
}
\description{
Super Tax function for use on HIDLA, as part of the Income Tax Assesment
For use with mapply()
}

