% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ysi_income_tax.R
\name{SUPERTAXfn}
\alias{SUPERTAXfn}
\title{Super Tax Function}
\usage{
SUPERTAXfn(Super_Inc = Super_Inc, Age = Age)
}
\arguments{
\item{Super_Inc}{Whole dollar value of super income}

\item{Age}{The age of the tax payer}
}
\value{
A vector
}
\description{
Super Tax function for use on HIDLA, as part of the Income Tax Assesment
}

