% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ysi_super_tax.R
\name{ysi_super_tax}
\alias{EmpSuperCont_fn_2}
\alias{ysi_super_tax}
\title{YSI Super Tax Assesment Function (Working...)}
\usage{
EmpSuperCont_fn_2(DoesEmpMakeCont = NULL, ContMadeByEmp = NULL,
  FreqEmpCont = NULL, Wage = NULL, EmpContPerWage = NULL)
}
\arguments{
\item{df}{Argument is intended to be a survey with the Tax_Matrix set, as a data frame}
}
\value{
A data frame
#' @export
}
\description{
This function combines all the nessesary Superannuation tax functions and data set mutations to be an all-in-one function for calculating Super Tax
(Requires Tax_Matrix set)
(Currently only usable in Wave14)
}

