% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_variables.R
\name{ysi_tax_inc}
\alias{ysi_tax_inc}
\title{YSI Taxable Income Function}
\usage{
ysi_tax_inc(df = NULL)
}
\arguments{
\item{df}{Argument is intended to be an edited HILDA data frame}
}
\value{
A data frame
}
\description{
This function mutates Taxable Income (TaxInc), a commonly used variable for YSI analysis as part of Income Tax Assesment
}
\details{
It generates taxable income as a variable. Dependent on having used:

ysi_mutate_combine()
DEDfn()
}

